+++ client/src/App.js
@@ -1,11 +1,21 @@
 import React, { Component } from 'react';
 import { MuiThemeProvider } from '@material-ui/core';
-import { BrowserRouter as Router } from 'react-router-dom';
+import { BrowserRouter as Router, Switch, Route, Link } from 'react-router-dom';
 import axios from 'axios';
 import { theme } from './themes/theme';
 import Navbar from './components/Navbar';
+import { createBrowserHistory } from 'history';
+
+import Home from './pages/Home';
+import List from './pages/List';
+import Messages from './pages/Messages';
+import Profile from './pages/Profile';
+import ProfileDetailForm from './components/ProfileDetailForm';
+import Jobs from './pages/Jobs';
+import Payment from './pages/Payment';
 
 // import "./App.css";
+export const history = createBrowserHistory();
 
 class App extends Component {
   constructor() {
@@ -39,11 +49,12 @@ class App extends Component {
       } else {
         this.setState({
           loggedIn: false,
-          userEmail: null,
+          userInfo: null,
         });
       }
     });
   }
+
   render() {
     return (
       <MuiThemeProvider theme={theme}>
@@ -53,7 +64,35 @@ class App extends Component {
             loggedIn={this.state.loggedIn}
             userInfo={this.state.userInfo}
           />
+
+          <Switch>
+            <Route exact path="/">
+              <Home />
+            </Route>
+            <Route exact path="/list">
+              <List />
+            </Route>
+            <Route exact path="/notifications">
+              <Jobs />
+            </Route>
+            <Route exact path="/jobs">
+              <Jobs />
+            </Route>
+            <Route path="/payment">
+              <Payment />
+            </Route>
+            <Route path="/messages">
+              <Messages />
+            </Route>
+            <Route path="/profile/edit">
+              <Profile />
+            </Route>
+            <Route path="/details" component={ProfileDetailForm} />
+            {/* </Route> */}
+          </Switch>
         </Router>
+
+        {/* <Route path="/" exact render={(props) => <LoginPage {...props} />} /> */}
       </MuiThemeProvider>
     );
   }
